requirements.txt
asgiref==3.2.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
boto3==1.9.219
botocore==1.12.219
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-cors-headers==3.1.0
django-haystack==2.8.1
django-npm==1.0.0
django-redis==4.10.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.7.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docutils==0.15.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
zope.interface==4.6.0
